Rename Primary Cells and Secondary Cells
What
---
1. Rename Primary Cell to `Legacy Cell`
2. Rename Secondary Cells to `Cells`

Why
---
In <https://gitlab.com/gitlab-com/content-sites/handbook/-/merge_requests/8284> we defined the existing GitLab.com installation as Legacy Cell.
The term Primary Cell was an old term before we introduced the Topology Service.

The terminology `Primary` and `Secondary` turned out to be confusing
because people might think that we have some kind of fallback mechanism
between the two.

During meetings, we have been calling it Legacy Cell all the time but we
never updated the documentation.