docs: Modify Rate Limiting design document
Based on all of the discussions over the past couple of weeks, both
synchronous and asynchronous in the MR, this has been used to create an
updated proposal of the design doc that instead of focusing on creating
a new service initially, instead focuses on standardising and improving
the basic configuration of our rate limiting before we look at adding an
interface layer in front of it.

Alongside this, managing IP based bypasses has been raised as a pain
point that is often neglected. I believe I've identified a way we should
be able to centralise this configuration in Cloudflare, and remove the
Bypass header logic from HAProxy completely which would be a huge
improvement on it's own.