ci: Run all code quality jobs on the default branch.
If the code quality reports are not run on the target branch, then GitLab
reports _all_ findings as "new" in merge requests. This reduces the value of
code quality findings significantly, because we teach users to ignore code
quality reports.

This changes all jobs with `codequality` type artifacts to run on the `main`
branch, too.